add: newPoint, newParenthesis, result (int or double)

+++ src/main/java/main/calculadora/Calculadora.java
@@ -3,17 +3,17 @@ package main.calculadora;
 import java.util.Stack;
 
 public class Calculadora {
-    public int performCalculation(String expresion) {
+    public Double performCalculation(String expresion) {
         // Dividir el texto en tokens (números y operadores)
         String[] tokens = expresion.split("(?<=\\d)(?=\\D)|(?<=\\D)(?=\\d)");
 
         // Pilas para números y operadores
-        Stack<Integer> numbers = new Stack<>();
+        Stack<Double> numbers = new Stack<>();
         Stack<Character> operators = new Stack<>();
 
         for (String token : tokens) {
             if (isNumber(token)) {
-                numbers.push(Integer.parseInt(token));
+                numbers.push(Double.parseDouble(token));
             } else if (isOperator(token)) {
                 char operator = token.charAt(0);
                 while (!operators.isEmpty() && precedence(operators.peek()) >= precedence(operator)) {
@@ -49,11 +49,11 @@ public class Calculadora {
     }
 
     // Realiza la operación correspondiente
-    private void performOperation(Stack<Integer> numbers, Stack<Character> operators) {
-        int num2 = numbers.pop();
-        int num1 = numbers.pop();
+    private void performOperation(Stack<Double> numbers, Stack<Character> operators) {
+        double num2 = numbers.pop();
+        double num1 = numbers.pop();
         char operator = operators.pop();
-        int result = switch (operator) {
+        double result = switch (operator) {
             case '+' -> num1 + num2;
             case '-' -> num1 - num2;
             case '*' -> num1 * num2;
